MSF-234 - Don't archive logs if they do not exist.

+++ AntiVirus/Scripts/logrotate.sh
@@ -4,10 +4,16 @@ logs=/var/log
 logFile=amavis.log
 amavisLog="${logs}/${logFile}"
 
-cp ${amavisLog} "${logs}/$(($(date +%s%N)/1000000))_$logFile"
-echo "Log initialized: `date`" > ${amavisLog}
+if test -e ${amavisLog}
+then
 
-find ${logs} -name "*_$logFile" -mtime +120 -exec rm -f {} \; >> ${amavisLog}
-echo "Beginning of log file:" >> ${amavisLog}
+  cp ${amavisLog} "${logs}/$(($(date +%s%N)/1000000))_$logFile"
+  echo "Log initialized: $(date)" > ${amavisLog}
+  find ${logs} -name "*_$logFile" -mtime +120 -exec rm -f {} \; >> ${amavisLog}
+  echo "Beginning of log file:" >> ${amavisLog}
+else
+
+  echo "Log not yet available for archiving: $(date)" >> ${amavisLog}
+fi
 
 sleep 604800; sh /logrotate.sh &
